[Task/CRA-101_rebase]: changed scored marks data type from int to double

+++ android/MaterialDesignTest/app/src/main/java/com/testexample/materialdesigntest/data/model/Result.kt
@@ -15,7 +15,7 @@ data class Result(
     val noOfQuestionsCorrect: Int,
     val totalNoOfQuestions: Int,
     val totalMarks: Int,
-    val scoredMarks: Int
+    val scoredMarks: Double
 ) : Parcelable
